cli/operator: extract startup orchestration into run()
Move the startup body (network/CL setup through operatorNode.Start) out of the StartNodeCmd Run closure into run(ctx, cfg, logger) error in a new run.go. Pure move: the in-function Fatals and the mid-stream resolveAndValidate call are kept as-is; the fail-fast reposition and the Fatal->return conversions are separate follow-up commits.

The thin Run keeps only the bootstrap stage (build data, config load, observability init and the top-level defers) and calls run(), Fataling once on its returned error. Observability-options building moves into a buildObservabilityOptions helper so the bootstrap closure stays thin.
